pyproject.toml
[project]
name = "prop-firm-pilot"
version = "0.1.0"
description = "Fully automated FX trading system for E8 Markets prop firm accounts"
requires-python = ">=3.10,<3.11"
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.1.0",
  "pyyaml>=6.0.1",
  "loguru>=0.7.2",
  "python-dotenv>=1.0.1",
  "duckdb>=1.1.3",
  "pandas>=2.0.3,<2.1",
  "numpy>=1.21.0,<1.24",
  "pyqlib==0.9.7",
  "dropbox>=12.0.2",
  "apscheduler>=3.10.4",
  "curl-cffi>=0.14.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=4.0",
  "ruff>=0.1.0",
  "respx>=0.21.0",
]

[project.scripts]
prop-firm-pilot = "src.main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.uv]
index-url = "https://pypi.org/simple"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
